pyproject.toml
[tool.poetry]
name = "chatbot-langgraph"
version = "0.1.0"
description = ""
authors = ["mielesantos <mielesnts@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
langchain-groq = "^0.2.1"
langchain = "^0.3.7"
langgraph = "^0.2.44"
openai = "^1.53.0"
groq = "^0.11.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
ruff = "^0.4.8"
httpx = "^0.27.0"

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .&& ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'streamlit run main.py'
pre_test = 'task lint'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"